+++ app/B2020+12/src/main/java/de/stefanlober/b2020/view/IView.kt
@@ -1,5 +1,5 @@
 package de.stefanlober.b2020.view
 
 interface IView {
-    fun update(dataList: ArrayList<ShortArray>)
+    fun update(dataList: ArrayList<DoubleArray>)
 }
